+++ src/components/Tips/TokenItem.tsx
@@ -1,7 +1,7 @@
 import { multipliedBy } from '@masknet/web3-shared-base';
 
 import { ClickableButton } from '@/components/ClickableButton.js';
-import { router, TipsRoutePath } from '@/components/Tips/tipsModalRouter.js';
+import { router, TipsRoutePath } from '@/components/Tips/TipsModalRouter.js';
 import { TokenIcon } from '@/components/Tips/TokenIcon.js';
 import { classNames } from '@/helpers/classNames.js';
 import { TipsContext } from '@/hooks/useTipsContext.js';
