+++ src/components/Tips/WalletSelector.tsx
@@ -5,7 +5,7 @@ import { Fragment, memo } from 'react';
 import ArrowDown from '@/assets/arrow-down.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { AddressLink } from '@/components/Tips/AddressLink.js';
-import { router, TipsRoutePath } from '@/components/Tips/tipsModalRouter.js';
+import { router, TipsRoutePath } from '@/components/Tips/TipsModalRouter.js';
 import { classNames } from '@/helpers/classNames.js';
 import { formatEthereumAddress } from '@/helpers/formatEthereumAddress.js';
 import { TipsContext, type TipsProfile } from '@/hooks/useTipsContext.js';
