+++ src/components/TokenProfile/CommunityLink.tsx
@@ -1,15 +1,15 @@
 import { LinkIcon } from '@heroicons/react/24/outline';
 import { skipToken, useQuery } from '@tanstack/react-query';
 
-import DiscordRound from '@/assets/DiscordRound.svg';
-import FacebookColored from '@/assets/FacebookColored.svg';
+import DiscordRound from '@/assets/discord-round.svg';
+import FacebookColored from '@/assets/facebook-colored.svg';
 import GitHub from '@/assets/GitHub.svg';
-import Instagram from '@/assets/Instagram.svg';
-import Medium from '@/assets/Medium.svg';
-import RedditRound from '@/assets/RedditRound.svg';
-import TelegramRound from '@/assets/TelegramRound.svg';
-import TwitterXRound from '@/assets/TwitterXRound.svg';
-import YouTube from '@/assets/YouTube.svg';
+import Instagram from '@/assets/instagram.svg';
+import Medium from '@/assets/medium.svg';
+import RedditRound from '@/assets/reddit-round.svg';
+import TelegramRound from '@/assets/telegram-round.svg';
+import XRound from '@/assets/x-round.svg';
+import YouTube from '@/assets/youtube.svg';
 import { Tooltip } from '@/components/Tooltip.js';
 import { SourceInURL } from '@/constants/enum.js';
 import { Link } from '@/esm/Link.js';
@@ -28,7 +28,7 @@ const brands: Record<CommunityType, React.ReactNode> = {
     medium: <Medium width={16} height={16} />,
     reddit: <RedditRound width={16} height={16} />,
     telegram: <TelegramRound width={16} height={16} />,
-    twitter: <TwitterXRound width={16} height={16} />,
+    twitter: <XRound width={16} height={16} />,
     youtube: <YouTube width={16} height={16} />,
     linkedin: <LinkIcon width={16} height={16} />,
     other: <LinkIcon width={16} height={16} />,
